+++ CHANGELOG.rst
@@ -2,11 +2,14 @@
 CHANGELOG
 =========
 
-1.1.6
-=====
+1.2
+===
 
+ * added support for Python 3.3
+ * dropped support for Python 2.6 (supported versions are 2.7 and 3.3)
+ * added runtime config option for running different Python versions on Gondor
+ * fixed newline output in gondor manage
  * fixed bug mercurial commit identification (use hg identify)
- * added runtime config option for running different Python versions
 
 1.1.5
 =====